Hide menu expansion by default
Currently, the menus start out fully expanded. When the javascript executes,
the unwanted menus are hidden. This happens very quickly, but you still see
a flicker at the beginning of the page load because of it. This PR hides
the entire menu until the javascript has executed. It is possible for it
to take a second for this to happen, but I think that's preferrable to
having the entire menu either flicker on every page load or be displayed
fully expanded until the JS executes.