generate: vendor generator deps
the generator script is pretty simple and only depends on a couple external libs.
given our experience in the main repo, no fancy submodules or glide here -- just
a folder with a few deps in it.

Note that we're not vendoring everything depended on, since obviously we depend on
cockroachdb/cockroach but want to use the currentversion checked out in GOPATH.