add builtin func's long-form info if present
currently only replace(...) has a its Info set, so that's the one to look at.

This also switches to wrapping the function signatures in <code> blocks, to distinguish them from the prose describing them, which required switching 'linkType' to return HTML.

HTMl <code> blocks, rather than backticks, are required to preserve the links.