teach sidebar about alternate URLs
This commit teaches the sidebar that the secure and insecure variants of
the deploy-to-XXX-cloud-provider guides are essentially the same page.
Now, the sidebar only links to the secure variant (i.e., the `url`
specified in the sidebar data) as before, but the insecure variant
(i.e., the `alternate_urls` in the sidebar data) will expand and
highlight the secure variant in the sidebar, too.