Surface core value props in start a cluster
These changes attempt to keep the Start a Cluster tutorial
quick and easy while introducing new users to what makes
CockroachDB unique, not just ease of deployment, but also
automated replication and survival.

I've also stripped out some details that I think are not
important to the majority of users getting up and running
for the first time, e.g., how to customize `start` flags
and the fact that you can use third-party monitoring.

Finally, I've update the Stop the Cluster step to account
for the fact that `cockroach quit` can't be used to stop
nodes once quorum is lost.

Fixes #1191
Fixes #1252