Answer more FAQs in the docs
As we identify more and more FAQs, we need a good way
to make answers to these questions easily accessible/searchable
in the docs. This PR offers one possible approach: Adding
an FAQs section to the sidenav, with categories of FAQs broken
into distinct pages. For example, in addition to the general
FAQ we've had for some time, I've added a page on SQL-specific
FAQs.