Use single quotes throughout ruby example
There isn't really a standard style for this in ruby as far as I'm
aware, so I arbitrarily chose standardizing on single quotes because I
like the colour they highlight to on the docs page better. There's one
use of double quotes left because it uses string interpolation (which
single quotes disallow).