don't use non-versioned latest tarball in install instructions
The cockroach-latest*.tgz tarballs have the unfortunate downside of not
including the version in either the filename or the extracted folder.
This was a convenient way to avoid updating the install instructions to
reflect the new filenames for every release. Now, though, we track the
latest release's name in _data/strings.yml, so we can use that to
automatically derive release filenames.

This is an important step in resolving cockroachdb/cockroach#14731.

Closes #1368.