update build-from-source instructions with new `make install` command
The build from source instructions refer to an outdated location for the
built cockroach binary. Update them to suggest the new `make install`
command, which was added to simplify exactly this step. Also add a note
with the correct path to the built binary, for users who have a moral
objection to installing it into /usr/local.

Shout out to @couchand for reporting this issue.

Fixes #1416.