move ci and docker configuration into repository
Move the configuration for CI, which used to leave in the TeamCity build
configuration, and the configuration for Docker, which used to live in
cockroachdb/docs-docker, into the repository. This allows us to
scope CI/Docker configuration changes to one branch and atomically
update the configurations when the branch merges.