replace all references to `gh-pages` with `master`
We no longer use GitHub pages directly; we run Jekyll ourselves and push
the built site to another repository. Stop using the `gh-pages` branch
so that GitHub stops trying to build this repository as a GitHub pages
site (and warning about how the redcarpet gem is no longer supported).

This completes the transition by replacing all references to `gh-pages`
with `master`.