Refactor the SHOW/SET documentation.
Both SHOW and SET are really single statements which act on the
session (and they are also handled using a single code path in
CockroachDB!), with the exception of SET TRANSACTION which acts on the
current transaction. This patch acknowledges the commonality and
merges the various documentation pages into two more general pages.