untangle and speed up sidebar rendering
Rendering the sidebar with YAML and Liquid is expensive. Liquid loops,
in particular, are quite slow, and the sidebar loops must be rerun for
every page to inject the "active" class where appropriate.

Since our sidebar already doesn't function without JavaScript, we can
avoid all this overhead by instead rendering the sidebar with JavaScript
on users' browsers. This gets our total build time down to under 3s. The
only downside is that the sidebar data is now JSON, which is a bit more
annoying to edit by hand.