Fix rendering problem by using html codes inside code tags
See https://github.com/vmg/redcarpet/issues/546 for more on how redcarpet cares about spaces. There's a lax_spacing setting that might be part of the problem or solution, but as is right now just remembering to add some extra space now and again is a good enough quick fix.