new generate option --max-workers to increase/decrease concurrency
Prior to this patch the generator would create 1 goroutine per
diagram, that would in turn all run the railroad generator. Since
the railroad generator is a Java program one would swamp their
system with many generators instances side by side. (Many more
than cores, typically: there are dozens of diagrams to generate).

Moreover, the railroad diagram generator is already multithreaded, so
there is little gain to be obtained by this concurrency.

This patch controls the number of parallel processes with a new
configuration flag `--max-workers` which now defaults to 1.  A user
can experiment increasing this value to get more parallelism but the
expectation should be that little performance is to be gained this
way unless one has more than a few dozen cores available.