Update v1.1 Docker Swarm instructions to use the init command
Also switch to using an `attachable` network which makes running one-off
containers easier and simplify/fix the docs as applicable.

We could also go a little further with the secure guide and remove the
client keys from each container now that we can connect to the cluster
with one-off containers, but it'll probably still come in handy in
debugging situations to have the keys in all the containers.