Fix copying of a few commands in 1.1 swarm docs
Using the copy-to-clipboard functionality was copying the dollar signs
along with everything else, breaking the ability to paste the commands
directly into my terminal. The 1.0 docs were already correct.

It's still a little awkward that this copies the comments along with the
command, but that doesn't break anything (most of the time, at least).