Update Mesosphere docs to reflect support for newer cockroach versions
I don't know whether to include the doc in v2.0 or not - you can't
actually run unstable releases using the DC/OS package we provide, but
is it weird if we don't have the doc for it? Will we forget to add it
later? Let me know what you'd like.