Enhance various statements and diagrams
**tl;dr**: fix confusion, better document joins, fix upsert/delete/insert/update.

Prior to this patch, the documentation was confusing "SELECT
statements" and "selection clauses". This confusion was most visible in the documentation for INSERT and UPSERT, which incorrectly specified that these statements take a selection clause as operand, whereas they support much more than that.

The documentation was also incorrectly listing set operations (EXCEPT/UNION etc) as operators of SELECT in particular, whereas they actually work also with any selection clause.

The root cause of this sorry state of affairs was caused by a single
misconception: there is really no such thing as a "SELECT statement"
in the SQL grammar; instead there are various relational expressions
and operators, and selection clauses are just a particular instance of relational expressions.

This patch enhances this situation as follows:

- The page about a "SELECT statement" is deleted. It is replaced by a
  page titled "Simple SELECT clause", to properly highlight it is
  merely an instance of a selection clause (albeit with a reminder
  that it happens to also be valid as standalone statement).

- To ensure the basic SELECT syntax remains the "main SELECT doc page"
  for newcomers, the top-level links in navigation for the `SELECT`
  keyword are changed to point to the page on the simple SELECT
  clause.

- A new page "Statement-Like Relational Expressions" is introduced,
  which explains properly what can become a subquery. The pages about
  other statements that require a statement-like query are modified to
  point to this new page.

- The difference between "table expressions", "selection clauses" and
  "statement-like queries" is repeatedly clarified in numerous places,
  to ensure that clarity is pervasively maintained throughout the
  docs.

Additionally, the syntax diagrams for all these constructs are updated to more accurately reflect reality. Some more documentation for the various `JOIN` operators is added, as well as for subqueries. The term "value expression" is removed in favor of the more correct "scalar expression".

To review this patch, the reviewer is invited to first proceed in a
holistic manner and navigate the result of the changes with the eyes
of a newcomer:

1. Land on the "Simple SELECT clause" page; ensure that it makes
   sense, including its simplified syntax diagram.

2. Check out the page on "Table Expressions", which documents what can
   fit in the FROM clause of a simple SELECT clause. Check that it
   makes sense, including the new syntax diagrams.

3. Check the new page on "Join Expressions". Check that it makes sense, including the new syntax diagrams and "performance best practices".

4. Check the docs for INSERT/UPDATE/UPSERT/DELETE and check they make
   sense, including the new syntax diagrams.

5. Navigate to the page on "Relational Expressions" and check that it
   makes sense, including the new syntax diagrams.

6. Review the new page "Limiting Query Results" linked from there.

7. Check The new page "Subqueries".

Once the reviewer satisfies themselves that the relationship between
the pages and the syntax diagrams make sense, the line-by-line diff
can be reviewed for further edits.