Revise the sidenav
These are some minor changes to add clarity to
the division and priority of content. The biggest
change is place the prometheus tutorial with
the monitoring and alerting overview under deployment.
This is really where it needs to be, as monitoring
is an essential part of production deployment.
The other main change is renaming "Manage" to
"Maintain" to clarify that the section contains
maintenance-type tasks like upgrading a cluster
and decommissioning nodes.

Post 2.0, we'll be reorganizing/dividing the content
more into guides/reference/concepts categories.