Copy v1.0 known join flag limitation over to v1.1 known limitations
I'd guess this got dropped from the known limitations because the --join
flags in our examples changed to include more than just the first node,
but the underlying issue still exists in clusters large enough that it
isn't guaranteed that a copy of the first range will be on nodes
included in the join flags.

It's been fixed for 2.0, but we should still include this in our 1.1
docs. I've tweaked the wording a bit to make more sense than the v1.0
wording verbatim would have.