Declare install instructions as HTML, not Markdown
This page is straight HTML--no Markdown in sight--and so the Markdown
parser gets confused and starts escaping quotes in JavaScript halfway
through. Give it an .html extension to get the Markdown parser out of
the picture.

Also unbreak the "Edit this Page" link by using page.path instead of
page.url, which remembers the original extension of the page.

Fix #2926.