Update --host and --advertise-host guidance in deployment tutorials
Since it's recommended to leave `--host` out, this PR removes `--host`
from the `cockroach start` commands in our deployment tutorials and adds
a section explaining that default and when and how to set `--host` and
`--advertise-host`. This PR does the same for `--locality`, since that
flag is important mostly for multi-datacenter deployments.

This is a general follow-up to https://github.com/cockroachdb/docs/pull/3163.