Make back-to-top button an explicit front-matter setting
Now that the page toc is always in the right margin, and users
can use the toc to jump around the page, the back-to-top button
is only helpful in cases where a page is long and does not have
a toc. In these cases, writers can explicitly set
back_to_top: true in the pages front-matter.