Get rid of unnecessary versioned includes for kubernetes performance
The entire reason for originally using the include was to avoid
duplicating the material. If we're going to be duplicating the material
anyway, there's no benefit to putting it in _includes files.

I'd be happy to keep this in a _includes file if we're still open to
versionless includes, but it sounds like we want to avoid them as a
general rule.

Cleans up after #3366.