More correctly document SERIAL.
SERIAL has always been a pseudo-type. It never possessed the
attributes that are required of real types (For example, it does not
have values of its own. It cannot be stored. It cannot be produced by
scalar expressions. It does not have comparison rules. It is not
independent, like other types, from the presence or absence of a
DEFAULT clause in the same column definition CREATE TABLE.)

The notion that it could be documented as a data type was erroneous, a
past misunderstanding.

Moreover, due to compatibility problems with PostgreSQL, is is not
impossible that the translation of SERIAL to `DEFAULT unique_rowid()`
be changed in the future to use SQL sequences instead; if/when that
occurs a new keyword may be added to preserve the `rowid()` syntax.

Therefore, it is desirable in the current doc to spell out `DEFAULT
unique_rowid()` throughout all the tutorials and examples to ensures
the table definitions are future-proof.