Render search bar on Safari
Safari inexplicably clips `position: fixed` elements to the boundaries
of their parent. This seems to violate the CSS spec--the only containing
box of a `position: fixed` element should be the viewport--but the
solution is to add some padding on the parent element so that the search
bar is not outside of its bounding box.

Also remove `height: fit-content`, which caused Safari to add about 50px
of whitespace beneath the search bar that was obscuring the navigation.
Not sure what the `height: fit-content` was for; things seem to render
fine without it.