Split install into OS-specific pages
Historically, our core Install CockroachDB page gets a high number of
"Was this page helpful? Yes" responses, but it also gets too many
"Was this page helpful? No" responses. For example, since 1/18, the
 page got 227 Yes votes and 72 No votes.

Because the current page covers all OSs, it's hard to glean where users
are having problems. Therefore, this PR:

- Splits the install instructions into 3 OS-specific pages. This makes
it possible to track No votes per OS. It also make it easier to link
users to instructions for specific OS install methods.
- Changes the JS on the original install-cockroachdb.html to detect
the user's OS and redirect to the appropriate OS-specific install
page. All internal links to the install page still point here so as
to save the user the effort of selecting the right OS.

The OS-specific pages still include the html for tracking installs
for marketing.

Fixes #4031.
Fixes #1856.