Split install into OS-specific pages
Historically, our core Install CockroachDB page gets a high number of
"Was this page helpful? Yes" responses, but it also gets too many
"Was this page helpful? No" responses. For example, since 1/18, the
 page got 227 yes votes and 72 no votes.

Because the current page covers all OSs, it's hard to glean where users
are having problems. This PR splits the install instructions into
3 OS-specific pages so we can track No votes per OS. It also makes it
easier to link users to instructions for specific OS install methods.

The pages still include the html for tracking installs for marketing.

The only functionality we lose is the ability to auto-select the right
OS instructions using JS, but I think that loss is minor compared to
the insight and improved maintainability we gain.

Fixes #4031.
Fixes #1856.