Change fn trait bound in rust transaction example
Previously, the trait bound on the `execute_txn` function was
`FnMut`, which allows the function to mutate external state
in the course of executing the transaction.  This is at odds
with the intended semantics of this retry loop, which expects
that the behavior is idempotent.

OTOH, there are valid use-cases for needing to mutate state,
for instance to keep track of the retry count.  But since the
purpose of this code sample is as an example for normal use,
this seems to be the right trait bound.