SQL Dump Known Issue
Documenting that dump won't work without manual edits when interleaved tables are present.

I'm hoping this will avoid some consternation for someone trying to restore down the road. Better to inform them if basic functionality won't work as expected, especially if the fix is easy. This could free up the bulk IO team to work on things without a workaround.