Make top-left-most link to Install page
After refreshing the docs landing page, release note sign-ups
dropped noticably. In the previous version, the link to the
Install page was top-left, which is the position that tends to
get the most clicks.

This PR bumps the Install link to the top-left again in the hope
that more users once again go to the install page and, from there,
sign up for release notes.

This PR also adds a link to our training curriclum under
Learn More.