Be louder about experimental status of windows build
This "Install CockroachDB on Windows" page has a warning box but the
bolded text is "Native CockroachDB on Windows requires Windows 8 or
higher" and the experimental status is a bit hidden in the rest of the
box. This binary is almost entirely untested, so I'd suggest we swap the
bolded text to be about the experimental status of the binary.
Otherwise, I worry that users may form a bad initial impression of our
performance/stability/maturity (and perhaps don't bother contacting us
as a result).