Make local build of standard docs faster
This PR adds _config_standard_local.yml and adjusts the Makefile
to call this config as part of `make standard`. The config
excludes the markdown files for all versions earlier than 2.1
and 19.1, dramatically cutting the number of files that get built.

Once we have 19.2 docs, we'll need to update this, and then again
when 19.2 becomes the stable version. There's probably a less
fragile way to do this, but this is a good first step.

Builds for our live site aren't affected, as they use the
standard-build and managed-build make targets. Local builds
of the managed docs also are not affected.

Fixes #4822.