Know and get latest release version of cocroachdb always during up-gradation.
Aware that the latest version of cockroachDB is v19.1.3, but got confused when it stated from "v2.0" to "v19.1" - which is the latest version as per the versioning standards. The release details helped me to identify the latest version in cockroachDB, hope this will help others in identifying the updated version.