Tweak syntax highlighting prompt detection
The terminal prompt was using the same class as double-quoted SQL
strings in the new version of the syntax highlighter, which was causing
the prompt detection code to inadvertently remove double-quoted SQL
strings.

Teach the prompt detection code to double-check that the prompt it's
removing is actually at the beginning of the code block and only
contains the prompt character ($) or (<) to avoid accidentally removing
strings.