Update 19.2 Python code with app retry loop
Also, delete the comments in the Python code that state that
`crdb_internal.force_retry()` must be run as root.  This restriction was
lifted in https://github.com/cockroachdb/cockroach/pull/39246.

The 19.1 Python code samples were updated in #5085 and #5173.  We then
forgot to port those changes to 19.2.  This commit remedies that
oversight.

Addresses part of #5176.