Fix copy/paste commands on install pages
After a combination of https://github.com/cockroachdb/docs/pull/5358 and
https://github.com/cockroachdb/docs/pull/5386, some prompts weren't showing,
and others weren't being removed when copied via the copy button.

This PR just updates the class information to leverage this CSS:
https://github.com/cockroachdb/docs/blob/master/css/customstyles.css#L1675.
We're not relying on this in other places, but it seems we need to on the
install pages right now, since they're raw html. There's plans to discusss
this page with Marketing, since the html format is a requirement for their
install tracking.

Fixes #5419.