fix prompt marker detection
Previously we would mark something as a prompt if it matched the
character and was preceeded by nothing or a string of only whitespace
characters. However with strings like `id > 0` the space before the `>`
would be seen as empty since the `id` was a non-text DOM node.

To fix this, require that the `>` be after a newline.

Fixes #5480