Update make commands, configs, and build files to reflect new naming
- "managed" -> "cockroachcloud"
- "standard" -> "cockroachdb"

This should work for the PR-based TeamCity configs for Docs,
GitHub CI and Nightly. However, I'm not clear on how things work
post-merge now. The TC config for that doesn't reference Netlify,
where our docs are now hosted. Have to investigate further to
understand what, if anything, needs to change for building the live
site.