Revisions
This commit revises the new performance overview page for clarity
and efficiency of structure and language. It also situates it
in a new "Deploy > Performance" section, for now, with benchmarking
instructions, performance best practices, and performance tuning.

The performance page is a solid start, but longer term, I think we
should continue clarifying and focusing the page. It's still too
much a mixture of details about benchmarks rather than statements
about our performance characteristics that can be demonstrated by
benchmarks.