Fix inaccurate TIMESTAMPTZ description
The timestamp docs specified that our TIMESTAMPTZ datatype stores a time
zone offset. This is not true. It always stores the timestamp in UTC.
This is explicitly stated further down on the same page: "However, it is
conceptually important to note that TIMESTAMPTZ does not store any time
zone data."

For some reason this was correct in the v2.0 docs and earlier but
incorrect in later docs.