Exclude version selector form CC docs
After fixing sitemap generation, the version selector on CC
docs was causing link breakages that we don't understand. The
author of the version selector is no longer at CL, unfortunately.
Long-term, we need to understand and fix the version selector,
but for now, I'm removing it from CC docs to unblock this.