Customize page metadata for CockroachCloud
Previously, the title and description tags on CC docs pages
were referencing CockroachDB, which we think was negatively
effective SEO. This PR adds logic to replace "CockroachDB"
with "CockroachCloud" in the meta title and desciption fields
on pages with the site.cockroachcloud variable, which gets
added to pages built with _config_cockroachcloud.yml.

Fixes #5876.