Tweak docs landing page
Add Performance page. Also add link to Cockroach U and SQL Tuning tutorial,
to keep first row symmetrical.

This is just a stop-gap. it doesn't bring the docs landing page in line
with the sidenav. The Docs team is planning an offsite to more
broadly rethink the docs info architecture, so we should wait until that
outcome is clear before more dramatically revamping the landing page.

Fixes #5839.