scripts: update airtable entry on issue create/update
This works by searching a github issue based on title, and updating if necessary.

I haven't run this on every record yet, and there's rate limiting from github which rate limits us to 20 searches per minute, but otherwise it seems to work for the one record I tried.

Also added `--dry_run` mode so that it doesn't make records but does all the processing we want.