teach make test to build
htmltest assumes the _site directory has a working site layout. When building with just jekyll (i.e., not using netlify/build or netlify/local), the cloud directory is not in the expected place, so things break.

Add this make target so people can just type "make test" which uses the docker builder to simulate exactly what happens on netlify, which htmltest can then check since things are in the expected locations.

Fixes #4093
Closes #6157