Update redirects to use correct YAML var name
According to the documentation for the Jekyll redirect
plugin (https://github.com/jekyll/jekyll-redirect-from) we are using,
the variable name to use in the YAML header is

  `redirect_from`

Prior to this change, we had used the variable name

  `redirect-from`

which is apparently not understood by the Jekyll plugin.

This was causing our redirects to break.

Note: Unfortunately, this redirection plugin does not work for anchors
to specific locations on the page, so that visiting e.g.,

https://www.cockroachlabs.com/docs/stable/create-security-certificates.html#create-the-certificate-and-key-pair-for-a-client

redirects to the top-level page URL

https://www.cockroachlabs.com/docs/v19.2/cockroach-cert.html

which means the user is not directed to the specific location implied by
the anchor.