Remove intro step and make tutorial take up more of screen
Still need to figure out how to remove

top-padding: 100px

from the .content class. That padding is there to
accomodate the version switch bar. This will be
going away with the redesign, but ahead of that,
I need to remove it from this tutorial page
because it's pushing the prompt in the terminal
below the fold.