Update links on main docs landing page
However, it's not yet clear how we'll differentiate the main
docs landing page from the crdb docs landing page. The file
I've edited I would expect to eventually be a crdb docs landing
page, not the main one.

Once we've figured out how this works,
we'll have to copy the content from the v20.1/index.md file
into the file for the main landing page and fix some of
the links. At that time, we'll also need to put in place
the links specific to the crdb docs landing page.