Include generated cluster setting docs since 2.1
Instead of copying the generated files from the cockroach
repo to the docs repo, we can now use the remote_include plugin
to pull directly from the cockroach repo.

For cluster settings, I could go all the way back to 2.1, but
generated cluster settings docs don't seem to be on release-2.0
and earlier.

This is the first part of #9883.