Use client-side redirect for collate docs
The functions and operators page is auto-generated from cockroach.
That page includes invalid links to collatedstring.html. Even with
a server-side redirect from collatedstring.html to collate.html,
the build fails because the link checker we run doesn't know about
the server-side redirects.

For the time-being, I'm putting the client-side redirects back in
place. I'll open a PR in the cockroach repo to fix the links
in the generated docs and then remove the client-side redirects.