sql: clarify that role options are not inherited (#11675)
This also includes a rewording to use the term "role option" instead of
"parameter" to describe these properties.

Also, I removed a bullet point that said that roles cannot login. That's
no longer true, and there's an example at the bottom of the page that
shows how to create a role that can login.