Ensure we reference a tagged release of the operator
We're going to be making some changes on the master branch of the
operator including some of the manifest files referenced here in the
docs. Before we do, I wanted to make sure that it wouldn't break
anything that we're recommending people do to use the operator.

I've added `operator_version` to the main site config and referenced
that in all of the _start-cockroachdb-operator-secure.md_ files. This
changes the links from `https://.../master/...` to
`https://.../<version>/...`. Now when we change something on master, we
don't need to worry about it breaking our documentation.

The one downside here is that we'll need to change this value after we
cut a new release. I think this is worth the tradeoff and can easily be
added to our deployment steps.