Single page for all 21.1 releases
This is an initial experiment to understand the UX pros/cons
of putting all release notes from a major release on
a single page.

One obvious challenge: Our anchor links don't work as a user
would expect when there are multiple identical headings on
a page. The anchor link always links to the first instance.
We would need to fix this.

We'd also need to make many changes to the way release notes
are referenced (e.g., from the release note list, install pages,
sidenav, etc.) and put in place many redirects. Time-consuming
but doable and worth-while if there's consensus that this format,
or something related, is better for users.

Addresses #10963