Update references for operator v2.2.0
We've made a few changes for the v2.2.0 release of the public operator
including the location of the pre-generated manifests. Both the CRD and
the operator yaml files now live in the _install_ directory on master.

This PR updates the version to v2.2.0 and updates the docs to point to
the new locations.