Document the change in the default namespace
With the next release of the public operator, we've moved away from
using the `default` namespace. This has been a requested feature and is
generally a best practice.

The new version (yet to be released) will use
`cockroach-operator-system` and support reconciliation of resources in
all namespaces. Namespaces can be limited by setting the
`WATCH_NAMESPACES` environment variable for the manager deployment.