Product area category per page via front-matter
This PR is an experiment to see how we might use a
front-matter field to more easily bucket pages
by product area for the purpose of page analytics.

Specifically, we set an href on the page based
on the category front-matter field and then use
JS to auto-click the link so it will get
tracked in Google Analytics as a default event.