Reorganizes the security docs and adds several new pages, finishing DOCS-2559, DOCS-422, DOCS-458, DOCS-958.
The primary purpose of the reorganization is to properly separate descriptive specification-type docs from procedural 'how to do X' type docs. Previously most of the security docs, descriptive and procedural, are all together under the 'manage' header.

This PR adds a navigation header: Reference / Security, under which there is now a new 'Security Reference Overview Page', which specifies security features of the core release and compares features supported by paid products.

The Reference / Security section now also contains new pages for conceptual 'domains' of security concern, which I have broken down into Authentication, Authorization, Network Security, Encryption, and Availability/Resilience (loosely corresponding to the STRIDE model of security threats). I have not added a reference page for the last of these, but I would like to, offering a descriptive breakdown of the features of CockroachDB which specifically contribute to resilience in the face of attacks and disruptions (the fundamentally distributed nature of the db), but this requires thought so I'm creating a separate ticket.
The content of most of these pages has been scooped out of corresponding docs under the Manage/ Security Section, but the Authentication and Network Security pages are new; the former contains mostly content pertaining to the authentication configuration (HBA) issues; the latter is a new page and is a bit of a place holder but can be filled out with more content (as well as links to relevant procedural docs).

The PR also updates references to TLS, as we now support TLS 1.2 or 1.3 for client-node authentication.