Cloud and region availability FAQ
Users often want to know what regions CockroachDB is available in. Add an FAQ telling them how to find this information. This commit intentionally does not list out the specific clouds and regions because we don't want to worry about updating the docs each time a new region is added.